Large update containing bugfixes and d lot of refactoring

+++ UamobiUtil/widgets/MultibranchWidgets/DocResultsWidget.cpp
@@ -2,6 +2,7 @@
 #include "widgets/utils/ElementsStyles.h"
 #include "widgets/ElementWidgets/ProcessingOverlay.h"
 #include "widgets/ExtendedDelegates/ZebraListItemDelegate.h"
+#include <QMessageBox>
 #ifdef DEBUG
 #include "debugtrace.h"
 #endif
@@ -22,7 +23,7 @@ void DocResultsWidget::setIndexation(XmlObjects& settings)
 			settings.first()->value("to")
 		);
 		bool ok = settings.first()->value("last") == "true";
-		if (ok)
+		if (settings.count() > 1 && settings.at(1)->hasField("allowdelete"))
 		{
 			deleteAllButton->show();
 			deleteSelectedButton->show();
@@ -33,6 +34,10 @@ void DocResultsWidget::setIndexation(XmlObjects& settings)
 			deleteSelectedButton->hide();
 		}
 		nextButton->setDisabled(ok);
+		if (settings.first()->value("from") == "0")
+		{
+			pagenumber = 0;
+		}
 		previousButton->setDisabled(pagenumber >= 0);
 	}
 }
@@ -47,8 +52,10 @@ DocResultsWidget::DocResultsWidget( QWidget* parent)
 	nextButton(new MegaIconButton(innerWidget)), itemInfoStorage(new QListView(innerWidget)),
 	footerLayout(new QHBoxLayout(innerWidget)), backButton(new MegaIconButton(innerWidget)),
 	saveButton(new MegaIconButton(innerWidget)), items(new DataEntityListModel(this)), pagenumber(0), 
-	awaiter(new RequestAwaiter(AppSettings->timeoutInt + 20000, this))
+	awaiter(new RequestAwaiter(AppSettings->timeoutInt + 20000, this)), localCache(),
+	attachedControls(new ControlListWidget(this))
 {
+	detrace_DCONSTR("DocResults");
 	untouchable = innerWidget;
 	main = this;
 	current = innerWidget;
@@ -71,10 +78,11 @@ DocResultsWidget::DocResultsWidget( QWidget* parent)
 	innerLayout->setSpacing(0);
     mainLayout->setContentsMargins(0,0,0,0);
     mainLayout->setSpacing(0);
+	mainLayout->addWidget(attachedControls);
+	attachedControls->hide();
 
 	QSizePolicy mi(QSizePolicy::Maximum, QSizePolicy::Maximum);
 	QSizePolicy ma(QSizePolicy::MinimumExpanding, QSizePolicy::Maximum);
-	QFont scf = makeFont(0.04);
 
 	previousButton->setSizePolicy(mi);
 	previousButton->setIcon(QIcon(":/res/prevpage.png"));
@@ -88,20 +96,22 @@ DocResultsWidget::DocResultsWidget( QWidget* parent)
 	nextButton->setStyleSheet(NAVIGATE_BUTTONS_STYLESHEET);
 	nextButton->setMinimumWidth(calculateAdaptiveWidth(0.2));
 
+
+
 	indexationInfo->setSizePolicy(ma);
-	indexationInfo->setFont(scf);
+	indexationInfo->setFont(GENERAL_FONT);
 	indexationInfo->setAlignment(Qt::AlignCenter);
     itemInfoStorage->setWordWrap(true);
     itemInfoStorage->setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
 	//TODO: icons
 
-	deleteAllButton->setFont(scf);
+	deleteAllButton->setFont(GENERAL_FONT);
 	deleteAllButton->setText(tr("delete all"));
 	deleteAllButton->setIcon(QIcon(":/res/data.png"));
 	deleteAllButton->setStyleSheet(DELETE_BUTTONS_STYLESHEET);
 	deleteAllButton->hide();
 	
-	deleteSelectedButton->setFont(scf);
+	deleteSelectedButton->setFont(GENERAL_FONT);
 	deleteSelectedButton->setText(tr("delete selected"));
 	deleteSelectedButton->setIcon(QIcon(":/res/deleteData.png"));
 	deleteSelectedButton->setStyleSheet(CANCEL_BUTTONS_STYLESHEET);
@@ -116,29 +126,39 @@ DocResultsWidget::DocResultsWidget( QWidget* parent)
 	saveButton->setStyleSheet(COMMIT_BUTTONS_STYLESHEET);
 
 	userInfo->setText(tr("doc_results_userinfo"));
-	userInfo->setFont(scf);
+	userInfo->setFont(GENERAL_FONT);
 	userInfo->setAlignment(Qt::AlignCenter);
 
-	itemInfoStorage->setFont(scf);
+	itemInfoStorage->setFont(GENERAL_FONT);
 
 	itemInfoStorage->setModel(items);
 	itemInfoStorage->setItemDelegate(new ZebraItemDelegate(this));
 #ifdef QT_VERSION5X
 	QObject::connect(deleteAllButton, &MegaIconButton::clicked, this, &DocResultsWidget::deleteAll);
-	QObject::connect(deleteSelectedButton, &MegaIconButton::clicked, this, &DocResultsWidget::deleteCurrent);
+	QObject::connect(deleteSelectedButton, &MegaIconButton::clicked, this, &DocResultsWidget::getAttachedControls);
 	QObject::connect(backButton, &MegaIconButton::clicked, this, &DocResultsWidget::backRequired);
 	QObject::connect(saveButton, &MegaIconButton::clicked, this, &DocResultsWidget::saveDocument);
 	QObject::connect(nextButton, &MegaIconButton::clicked, this, &DocResultsWidget::nextPage);
+	QObject::connect(attachedControls, &ControlListWidget::controlsConfirmed, this, &DocResultsWidget::attachedControlsDone);
+	QObject::connect(attachedControls, &ControlListWidget::backRequired, this, &DocResultsWidget::hideCurrent);
 	QObject::connect(previousButton, &MegaIconButton::clicked, this, &DocResultsWidget::previousPage);
 	QObject::connect(awaiter, &RequestAwaiter::requestTimeout, this, &DocResultsWidget::was_timeout);
+	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::items_response);
+	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::save_response);
+	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::get_attached_response);
 #else
     QObject::connect(deleteAllButton, SIGNAL(clicked()), this, SLOT(deleteAll()));
-    QObject::connect(deleteSelectedButton, SIGNAL(clicked()), this, SLOT(deleteCurrent()));
+    QObject::connect(deleteSelectedButton, SIGNAL(clicked()), this, SLOT(getAttachedControls()));
 	QObject::connect(backButton, SIGNAL(clicked()), this, SIGNAL(backRequired()));
 	QObject::connect(saveButton, SIGNAL(clicked()), this, SLOT(saveDocument()));
 	QObject::connect(nextButton, SIGNAL(clicked()), this, SLOT(nextPage()));
+	QObject::connect(attachedControls, SIGNAL(controlsConfirmed()), this, SLOT(attachedControlsDone()));
+	QObject::connect(attachedControls, SIGNAL(backRequired()), this, SLOT(hideCurrent()));
 	QObject::connect(previousButton, SIGNAL(clicked()), this, SLOT(previousPage()));
     QObject::connect(awaiter, SIGNAL(requestTimeout()), this, SLOT(was_timeout()));
+	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(items_response()));
+	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(save_response()));
+	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(get_attached_response()));
 #endif
 }
 
@@ -146,17 +166,14 @@ void DocResultsWidget::loadItems()
 {
 	if (awaiter->isAwaiting())
 		return;
-#ifdef QT_VERSION5X
-	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::items_response);
-#else
-	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(items_response()));
-#endif
+
 	if (localCache.contains(documentGetResults))
 	{
 		AppNetwork->execQueryByTemplate(localCache[documentGetResults], QString::number(pagenumber), awaiter);
 	}
 	else
 		FullItemEntity::sendGetRequest(pagenumber, awaiter);
+	awaiter->deliverResultTo(documentGetResults);
 	deleteAllButton->hide();
 	deleteSelectedButton->hide();
 	showProcessingOverlay();
@@ -168,17 +185,19 @@ void DocResultsWidget::refresh()
 	PolyResponse response = RequestParser::parseResponse(parser, RecEntity(new FullItemEntity()));
 	if (response.isError)
 	{
-		userInfo->setText(response.errtext);
+		QMessageBox::critical(this, tr("Error!"), response.errtext, QMessageBox::Ok);
+#ifdef DEBUG
+		detrace_NRESPERR(response.errtext);
+#endif
 	}
 	else
 	{
-		setIndexation(response.additionalObjects);
+		PolyResponse invoices = RequestParser::parseResponse(response, RecEntity(new InvoiceEntity()));
+		setIndexation(invoices.additionalObjects);
 		saveButton->setDisabled(response.objects.isEmpty());
-		items->setData(response.objects);
+		items->setData(invoices.objects + response.objects);
 	}
 }
-
-
 
 void DocResultsWidget::previousPage()
 {
@@ -202,37 +221,125 @@ void DocResultsWidget::saveDocument()
 {
 	if (awaiter->isAwaiting())
 		return;
-#ifdef QT_VERSION5X
-	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::save_response);
-#else
-	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(save_response()));
-#endif
+
 	if (localCache.contains(unlockDocument))
 		AppNetwork->execQueryByTemplate(localCache[unlockDocument],"true", awaiter);
 	else
 		AppNetwork->execQueryByTemplate(QueryTemplates::unlockDocument, "true", awaiter);
+	awaiter->deliverResultTo(unlockDocument);
 	showProcessingOverlay();
+}
+
+void DocResultsWidget::getAttachedControls()
+{
+	if (awaiter->isAwaiting())
+		return;
+	if (!itemInfoStorage->currentIndex().isValid())
+		return;
+	RecEntity currentItem = itemInfoStorage->currentIndex().data(DataEntityListModel::DirectAccess
+	).value<RecEntity>();
+	if (!localCache.contains(documentResultGetBox))
+	{
+		attachedControls->clearControls();
+		attachedControls->emplaceControl(
+			InputControl(
+				new InputControlEntity(
+					"qty", "Int", "0"))));
+		_hideAny(attachedControls);
+	}
+	else
+	{
+		QStringList temp;
+		temp << currentItem->getId() << currentItem->getTitle() << QString::number(currentItem->getAttachedNumber());
+		AppNetwork->execQueryByTemplate(localCache[documentResultGetBox],3, temp, awaiter);
+		awaiter->deliverResultTo(documentResultGetBox);
+		showProcessingOverlay();
+	}
 }
 
 void DocResultsWidget::items_response()
 {
+	if (awaiter->deliverHere(unlockDocument))
+		return;
 	refresh();
-	QObject::disconnect(awaiter, SIGNAL(requestReceived()), 0, 0);
 	hideProcessingOverlay();
 }
 
 void DocResultsWidget::save_response()
 {
-	if (awaiter->restext.contains("_"))
+	if (!awaiter->deliverHere(unlockDocument))
+		return;
+	if (awaiter->restext.startsWith('_'))
+	{
+		if (awaiter->restext.count() > 1)
+		{
+#ifdef DEBUG
+			debugtrace::getObject() << detr_supply::netErrorPossible <<
+				" debug info came from server during document save operation |"
+				<< awaiter->restext << '\n';
+#endif
+		}
 		emit done(RecEntity());
-	QObject::disconnect(awaiter, SIGNAL(requestReceived()), 0, 0);
+	}
+	QDomDocument doc;
+	doc.setContent(awaiter->restext);
+	if (isError(doc))
+	{
+		userInfo->setText(makeError(doc));
+	}
+	hideProcessingOverlay();
+}
+
+void DocResultsWidget::get_attached_response()
+{
+	if (!awaiter->deliverHere(documentResultGetBox))
+		return;
+	ResponseParser parser(new LinearListParser(awaiter->restext, awaiter->errtext));
+	NetRequestResponse<InputControlEntity> response 
+		= RequestParser::parseResponse<InputControlEntity>(parser);
+	if (response.isError)
+	{
+		QMessageBox::critical(this, tr("Error!"), response.errtext, QMessageBox::Ok);
+#ifdef DEBUG
+		detrace_NRESPERR(response.errtext);
+#endif
+	}
+	else
+	{
+		attachedControls->clearControls();
+		attachedControls->useControls(response.objects);
+		_hideAny(attachedControls);
+	}
+	hideProcessingOverlay();
+}
+
+void DocResultsWidget::attachedControlsDone()
+{
+	if (awaiter->isAwaiting())
+		return;
+	if (!itemInfoStorage->currentIndex().isValid())
+	{
+		_hideAny(innerWidget);
+		return;
+	}
+	RecEntity currentItem = itemInfoStorage->currentIndex().data(DataEntityListModel::DirectAccess
+	).value<RecEntity>();
+	if (!localCache.contains(docDeleteByBarcode))
+	{
+		AppNetwork->execQueryByTemplate(docDeleteByBarcode, currentItem->getId(), attachedControls->mappedControls(), awaiter);
+	}
+	else
+	{
+		AppNetwork->execQueryByTemplate(localCache[docDeleteByBarcode],
+			currentItem->getId(), attachedControls->mappedControls(), awaiter);
+	}
+	_hideAny(innerWidget);
 	hideProcessingOverlay();
 }
 
 void DocResultsWidget::was_timeout()
 {
 	userInfo->setText(tr("doc_results_timeout") + QString::number(AppSettings->timeoutInt));
-	QObject::disconnect(awaiter, SIGNAL(requestReceived()), 0, 0);
 	hideProcessingOverlay();
 }
 
@@ -240,56 +347,38 @@ void DocResultsWidget::deleteAll()
 {
 	if (awaiter->isAwaiting())
 		return;
-#ifdef QT_VERSION5X
-	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::items_response);
-#else
-	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(items_response()));
-#endif
+	int result =  QMessageBox::question(this, tr("Delete all"), tr("Delete all data?"), QMessageBox::Ok, QMessageBox::Cancel);
+	if (result != QMessageBox::Ok)
+		return;
 	if (localCache.contains(documentDeleteAll))
 		AppNetwork->execQueryByTemplate(localCache[documentDeleteAll], awaiter);
 	else
 		AppNetwork->execQueryByTemplate(QueryTemplates::documentDeleteAll, awaiter);
+	awaiter->deliverResultTo(documentDeleteAll);
 	deleteAllButton->hide();
 	deleteSelectedButton->hide();
 	showProcessingOverlay();
 }
 
-void DocResultsWidget::deleteCurrent()
+
+
+void DocResultsWidget::hideCurrent()
 {
-	if (awaiter->isAwaiting())
-		return;
-	if (!itemInfoStorage->currentIndex().isValid())
-		return;
-#ifdef QT_VERSION5X
-	QObject::connect(awaiter, &RequestAwaiter::requestReceived, this, &DocResultsWidget::items_response);
-#else
-	QObject::connect(awaiter, SIGNAL(requestReceived()), this, SLOT(items_response()));
-#endif
-	FullItem item = upcastRecord<FullItemEntity>(
-		itemInfoStorage->currentIndex().data(
-			DataEntityListModel::DirectAccess
-		).value<RecEntity>());
-	if (item.isNull())
-		return;
-	if (localCache.contains(docDeleteByBarcode))
-		item->sendDeleteThisRequest(localCache[docDeleteByBarcode], awaiter);
-	else
-		item->sendDeleteThisRequest(awaiter);
-	deleteAllButton->hide();
-	deleteSelectedButton->hide();
-	showProcessingOverlay();
+	_hideAny(innerWidget);
 }
 
 void DocResultsWidget::_makeOverloads(const QVector<QueryTemplates::OverloadableQuery>& overloads)
 {
 	using namespace QueryTemplates;
 	switch (
-		((overloads.count()>4) ? 4 : overloads.count())
+		((overloads.count()>5) ? 5 : overloads.count())
 		)
 	{
-	case 4:
-		localCache.insert(documentDeleteAll, overloads.at(3).assertedAndMappedCopy(
+	case 5:
+		localCache.insert(documentDeleteAll, overloads.at(4).assertedAndMappedCopy(
 		documentDeleteAll));
+	case 4:
+		localCache.insert(documentResultGetBox, overloads.at(3));
 	case 3:
     {
         QStringList t;
